Added locks to com port initialization and change; fixes corner case where multiple com-port changes could create multiple contending threads
Added _ready function which allows a program to more cleanly know when the
CP4PC XBee sub-system is ready.  I don't really like this breaking of
encapsulation but it makes some of the code cleaner rather than having to
adding in while loops around DDO calls, etc. to detect when everything is
ready to go.

Robustified and restructured the read_messages() parser and API_Message
extract() routing.  Split read_messages() into two parts (read_messages()
and process_message()).  Cases were observed on OSX where the rx_buffer
could contain incomplete messages.  The data processing pathway has been
modified to eliminate characters up until the first ~.  If there is
enough data in the buffer to extract but packet extracted was not valid
the parser will move on (where before it would continue to try and process
the same buffer over and over).  This allows the system to move on with
life when things get a little weird (which is rare, but possible).