Added ability to specify location of default-settings file.
This change is necessary because when scripts are compiled to an executable using py2exe, and the option to create a library zip file is selected, scripts are unable to read static files residing in that zip file - and so simulator_settings is unable to read settings_default.json.
This change allows you to manually specify the location of the json file from which to load default settings.